JSFIDDLE Graphs
Gwen added the starting point for her bar graph.  I am having trouble removing the second variable from the graph, and am unsure if we are going to be able to include the level of interactivity that we wanted (i.e. having the "top items" show up when you hover over there retailer)